Refactor composition example and input modules.
