+++ examples/composition/README.md
@@ -19,8 +19,9 @@ curl --data-binary "@test-events.ndjson" "http://localhost:8000/events"
 ```
 
 Logs of interest will be visible on the first terminal, and they will
-refer to three instances of CDP: `initial`, `middle` and `terminal`
-(the name of the respective services in `docker-compose.yml`).
+refer to four instances of CDP: `first`, `second`, `third` and
+`fourth` (the name of the respective services in
+`docker-compose.yml`).
 
 To clean up, cancel the process in the first terminal with Ctrl-C and
 then bring the stack down including volumes:
@@ -31,21 +32,25 @@ docker compose down -v
 
 ## What does it do?
 
-The example shows three pipelines that take input from an [HTTP
-input](/../../#http) and [`tail` input](/../../#tail) and execute some
-forwarding steps with the [`send-http`](/../../#send-http) and
-[`send-file`](/../../#send-file) functions. Also, events are logged in
-each step showing the pipelines they've been in.
+The example shows four pipelines that take input from an [HTTP
+input](/../../#http), [`tail` input](/../../#tail) and [`poll`
+input](/../../#poll) and execute some forwarding steps with the
+[`send-http`](/../../#send-http), [`send-file`](/../../#send-file) and
+[`expose-http`](/../../#expose-http) functions. Also, events are
+logged in each step showing the pipelines they've been in.
 
 ## Pipeline files
 
 The pipeline files are named after the docker compose services:
 
-- [`pipeline-initial.yaml`](pipeline-initial.yaml) is the pipeline
+- [`pipeline-first.yaml`](pipeline-first.yaml) is the pipeline
   that receives initial events via HTTP and forwards them to the
-  `middle` one via a file.
-- [`pipline-middle.yaml`](pipeline-middle.yaml) is the pipeline that
-  receives events from the `initial` one via a file and forwards them
-  to the `terminal` one via HTTP.
-- [`pipeline-terminal.yaml`](pipeline-terminal.yaml) is the pipeline
-  that receives events from the `middle` one via HTTP.
+  `second` one via a file.
+- [`pipline-second.yaml`](pipeline-second.yaml) is the pipeline that
+  receives events from the `first` one via a file and forwards them
+  to the `third` one via HTTP.
+- [`pipeline-third.yaml`](pipeline-third.yaml) is the pipeline that
+  receives events from the `second` one via HTTP and forwards them to
+  the `fourth` one via HTTP exposition.
+- [`pipeline-fourth.yaml`](pipeline-fourth.yaml) is the pipeline that
+  fetches events from the `third` pipeline via HTTP polling.
